feat: 라벨 리스트 렌더링시 알맞은 폰트색 적용
- getFontColor 함수를 두 파일에서 사용해서 util로 분리
- 레이블 리스트 렌더링시 함수를 사용하여 폰트색 정함
